requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.17.0
alpaca-py>=0.43.2
python-dotenv>=1.2.1
matplotlib>=3.10.7
beautifulsoup4>=4.12.3
selenium>=4.38.0
pandas-market-calendars>=5.1.3 